+++ tests/integration/proxy/v1/test_service.py
@@ -32,6 +32,7 @@ class ServiceTestCase(IntegrationTestCase):
             {
                 "sid": "KSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "chat_instance_sid": "ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "unique_name": "unique_name",
                 "default_ttl": 3600,
                 "callback_url": "http://www.example.com",
@@ -110,6 +111,7 @@ class ServiceTestCase(IntegrationTestCase):
             {
                 "sid": "KSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "chat_instance_sid": "ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "unique_name": "unique_name",
                 "default_ttl": 3600,
                 "callback_url": "http://www.example.com",
@@ -172,6 +174,7 @@ class ServiceTestCase(IntegrationTestCase):
             {
                 "sid": "KSaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "chat_instance_sid": "ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "unique_name": "unique_name",
                 "default_ttl": 3600,
                 "callback_url": "http://www.example.com",