+++ twilio/rest/api/v2010/__init__.py
@@ -82,6 +82,13 @@ class V2010(Version):
         return self.account.available_phone_numbers
 
     @property
+    def balance(self):
+        """
+        :rtype: twilio.rest.api.v2010.account.balance.BalanceList
+        """
+        return self.account.balance
+
+    @property
     def calls(self):
         """
         :rtype: twilio.rest.api.v2010.account.call.CallList